requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0
opencv-python==4.8.1.78
tensorflow==2.15.0
torch==2.1.1
transformers==4.35.2
requests==2.31.0
python-dotenv==1.0.0